Fix | `credentials` commands (#66)
* Extract argument default as a constant

* Provide custom config when rendering single file

* Parse AWS access keys files as csv

* Allow paths relative to `$HOME` for credentials

* Restrict profiles used to get organization
Only use `management` or `bootstrap` profiles to get organization accounts

* Handle missing configuration files
Project information required to set up credentials needs a
fairly complex logic to circunvent the possibility of missing configuration
files. If the information is not found on the `project.yaml` file or
the `build.env` file, the Terraform definition will be used as source of
truth, as a last resource the user will be prompted.